requirements.txt
pandas==1.5.3
 torchvision==0.15.2
 numpy==1.24.3
 torch==2.0.1
 scikit-learn==1.2.2
 matplotlib==3.7.1
 streamlit==1.30.0
 fastapi==0.100.0
 uvicorn==0.24.0
 neo4j==5.8.0
 redis==4.5.1

 sqlalchemy==2.0.19
 python-multipart==0.0.6
 pydantic==2.1.1
 transformers==4.31.0
 faiss-cpu==1.7.4
 python-dotenv==1.0.0
 Flask==2.3.3
 flask-cors==4.0.0